requirements.txt
Flask~=1.1.2
flask-login
apscheduler
pymongo
flask-mongoengine
python-decouple
pytest
pytest-cov